+++ models/demo/src/migration.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-import core, { TxOperations } from '@anticrm/core'
+import core, { AccountRole, TxOperations } from '@anticrm/core'
 import { MigrateOperation, MigrationClient, MigrationUpgradeClient } from '@anticrm/model'
 import contact, { EmployeeAccount } from '@anticrm/contact'
 import recruit from '@anticrm/model-recruit'
@@ -61,7 +61,8 @@ export const demoOperation: MigrateOperation = {
       await tx.createDoc<EmployeeAccount>(contact.class.EmployeeAccount, core.space.Model, {
         email: 'rosamund@hc.engineering',
         employee,
-        name: 'Chen,Rosamund'
+        name: 'Chen,Rosamund',
+        role: AccountRole.Owner
       })
     }